+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,20 @@
 # sit-onyx
 
+## 1.0.0-alpha.10
+
+### Major Changes
+
+- a4df4af: refactor: update CSS variables
+
+  | Old                   |                         New |
+  | --------------------- | --------------------------: |
+  | `--onyx-spacing-4xs`  |        `--onyx-spacing-5xs` |
+  | `--onyx-spacing-3xs`  |        `--onyx-spacing-4xs` |
+  | `--onyx-spacing-2xs`  |        `--onyx-spacing-3xs` |
+  | `--onyx-spacing-xs`   |        `--onyx-spacing-2xs` |
+  | `--onyx-color-text-*` | `--onyx-color-text-icons-*` |
+  | `--onyx-color-icon-*` | `--onyx-color-text-icons-*` |
+
 ## 0.1.0-alpha.9
 
 ### Minor Changes